Scrap details only when the second half tab is available
This really fixes the bug: it turned out that when there was an overtime
in the first half, time condition was fulfilled for particular event but
second half tab was unavailable.
What should happen in such cases is that details should not be scraped
and info like 'second half not available' should be reported.